ride_management/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
django-filter==23.5
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
pytest-django==4.7.0
django-cors-headers==4.3.1